+++ CHANGES.rst
@@ -36,6 +36,7 @@ master (in development)
     gracefully [GH-434]
   - Restore mouse support in the error list [GH-468]
   - `less` now correctly resolves relative paths in `data-uri` [GH-471]
+  - `go-errcheck` now properly uses package names as syntax checker arguments
 
 0.20 (Aug 12, 2014)
 -------------------